MAINT: Merged tests into a single file
