+++ scipy/optimize/tests/test_least_squares.py
@@ -3,12 +3,171 @@ import warnings
 
 import numpy as np
 from numpy.testing import (run_module_suite, assert_, assert_allclose,
-                           assert_raises, assert_equal, assert_almost_equal)
+                           assert_raises, assert_equal)
 
 from scipy.optimize import least_squares
 from scipy.sparse import issparse, lil_matrix
 from scipy.sparse.linalg import aslinearoperator
-from scipy.optimize._lsq_common import make_strictly_feasible
+from scipy.optimize._lsq_common import (
+    prepare_bounds, in_bounds, step_size_to_bound, find_active_constraints,
+    make_strictly_feasible, scaling_vector, intersect_trust_region)
+
+
+# Tests of functions from ._lsq_common.
+
+
+class TestBounds(object):
+    def test_prepare_bounds(self):
+        bounds = (
+            1.0,
+            [2.0, 3.0, 4.0]
+        )
+        x0 = np.zeros(3)
+        lb, ub = prepare_bounds(bounds, x0)
+        assert_equal(lb, np.ones(3))
+        assert_equal(ub, np.array([2.0, 3.0, 4.0]))
+
+    def test_in_bounds(self):
+        lb = -np.ones(3)
+        ub = np.ones(3)
+        x1 = np.array([-0.1, 1.0, 0.0])
+        x2 = np.array([1.5, 0.1, -0.5])
+        assert_(in_bounds(x1, lb, ub))
+        assert_(not in_bounds(x2, lb, ub))
+
+    def test_step_size_to_bounds(self):
+        lb = np.array([-1.0, 2.5, 10.0])
+        ub = np.array([1.0, 5.0, 100.0])
+        x = np.array([0.0, 2.5, 12.0])
+
+        s = np.array([0.1, 0.0, 0.0])
+        step, hits = step_size_to_bound(x, s, lb, ub)
+        assert_equal(step, 10)
+        assert_equal(hits, [1, 0, 0])
+
+        s = np.array([0.01, 0.05, -1.0])
+        step, hits = step_size_to_bound(x, s, lb, ub)
+        assert_equal(step, 2)
+        assert_equal(hits, [0, 0, -1])
+
+        s = np.array([10.0, -0.0001, 100.0])
+        step, hits = step_size_to_bound(x, s, lb, ub)
+        assert_equal(step, np.array(-0))
+        assert_equal(hits, [0, -1, 0])
+
+        s = np.array([1.0, 0.5, -2.0])
+        step, hits = step_size_to_bound(x, s, lb, ub)
+        assert_equal(step, 1.0)
+        assert_equal(hits, [1, 0, -1])
+
+        s = np.zeros(3)
+        step, hits = step_size_to_bound(x, s, lb, ub)
+        assert_equal(step, np.inf)
+        assert_equal(hits, [0, 0, 0])
+
+    def test_find_active_constraints(self):
+        lb = np.array([0.0, -10.0, 1.0])
+        ub = np.array([1.0, 0.0, 100.0])
+
+        x = np.array([0.5, -5.0, 2.0])
+        active = find_active_constraints(x, lb, ub)
+        assert_equal(active, [0, 0, 0])
+
+        x = np.array([0.0, 0.0, 10.0])
+        active = find_active_constraints(x, lb, ub)
+        assert_equal(active, [-1, 1, 0])
+
+        active = find_active_constraints(x, lb, ub, rtol=0)
+        assert_equal(active, [-1, 1, 0])
+
+        x = np.array([1e-9, -1e-8, 100 - 1e-9])
+        active = find_active_constraints(x, lb, ub)
+        assert_equal(active, [0, 0, 1])
+
+        active = find_active_constraints(x, lb, ub, rtol=1.5e-9)
+        assert_equal(active, [-1, 0, 1])
+
+        lb = np.array([1.0, -np.inf, -np.inf])
+        ub = np.array([np.inf, 10.0, np.inf])
+
+        x = np.ones(3)
+        active = find_active_constraints(x, lb, ub)
+        assert_equal(active, [-1, 0, 0])
+
+        # Handles out-of-bound cases.
+        x = np.array([0.0, 11.0, 0.0])
+        active = find_active_constraints(x, lb, ub)
+        assert_equal(active, [-1, 1, 0])
+
+        active = find_active_constraints(x, lb, ub, rtol=0)
+        assert_equal(active, [-1, 1, 0])
+
+    def test_make_strictly_feasible(self):
+        lb = np.array([-0.5, -0.8, 2.0])
+        ub = np.array([0.8, 1.0, 3.0])
+
+        x = np.array([-0.5, 0.0, 2 + 1e-10])
+
+        x_new = make_strictly_feasible(x, lb, ub, rstep=0)
+        assert_(x_new[0] > -0.5)
+        assert_equal(x_new[1:], x[1:])
+
+        x_new = make_strictly_feasible(x, lb, ub, rstep=1e-4)
+        assert_equal(x_new, [-0.5 + 1e-4, 0.0, 2 * (1 + 1e-4)])
+
+        x = np.array([-0.5, -1, 3.1])
+        x_new = make_strictly_feasible(x, lb, ub)
+        assert_(np.all((x_new >= lb) & (x_new <= ub)))
+
+        x_new = make_strictly_feasible(x, lb, ub, rstep=0)
+        assert_(np.all((x_new >= lb) & (x_new <= ub)))
+
+        lb = np.array([-1, 100.0])
+        ub = np.array([1, 100.0 + 1e-10])
+        x = np.array([0, 100.0])
+        x_new = make_strictly_feasible(x, lb, ub, rstep=1e-8)
+        assert_equal(x_new, [0, 100.0 + 0.5e-10])
+
+    def test_scaling_vector(self):
+        lb = np.array([-np.inf, -5.0, 1.0, -np.inf])
+        ub = np.array([1.0, np.inf, 10.0, np.inf])
+        x = np.array([0.5, 2.0, 5.0, 0.0])
+        g = np.array([1.0, 0.1, -10.0, 0.0])
+        v, dv = scaling_vector(x, g, lb, ub)
+        assert_equal(v, [1.0, 7.0, 5.0, 1.0])
+        assert_equal(dv, [0.0, 1.0, -1.0, 0.0])
+
+
+class TestTrustRegion(object):
+    def test_intersect(self):
+        Delta = 1.0
+
+        x = np.zeros(3)
+        s = np.array([1.0, 0.0, 0.0])
+        t_neg, t_pos = intersect_trust_region(x, s, Delta)
+        assert_equal(t_neg, -1)
+        assert_equal(t_pos, 1)
+
+        s = np.array([-1.0, 1.0, -1.0])
+        t_neg, t_pos = intersect_trust_region(x, s, Delta)
+        assert_allclose(t_neg, -3**-0.5)
+        assert_allclose(t_pos, 3**-0.5)
+
+        x = np.array([0.5, -0.5, 0])
+        s = np.array([0, 0, 1.0])
+        t_neg, t_pos = intersect_trust_region(x, s, Delta)
+        assert_allclose(t_neg, -2**-0.5)
+        assert_allclose(t_pos, 2**-0.5)
+
+        x = np.ones(3)
+        assert_raises(ValueError, intersect_trust_region, x, s, Delta)
+
+        x = np.zeros(3)
+        s = np.zeros(3)
+        assert_raises(ValueError, intersect_trust_region, x, s, Delta)
+
+
+# Tests of least_squares.
 
 
 def fun_trivial(x, a=0):
@@ -50,7 +209,7 @@ def fun_rosenbrock_cropped(x):
     return fun_rosenbrock(x)[0]
 
 
-# When x 1-d array, return is 2-d array.
+# When x is 1-d array, return is 2-d array.
 def fun_wrong_dimensions(x):
     return np.array([x, x**2, x**3])
 
@@ -91,7 +250,7 @@ class BroydenTridiagonal(object):
             self.sparsity = None
             self.jac = lambda x: self._jac(x).toarray()
         else:
-            assert False
+            assert_(False)
 
     def fun(self, x):
         f = (3 - x) * x + 1
@@ -119,7 +278,6 @@ class BaseMixin(object):
 
     def test_args_kwargs(self):
         # Test that args and kwargs are passed correctly to the functions.
-        # And that kwargs are not supported by 'lm'.
         a = 3.0
         for jac in ['2-point', '3-point', 'cs', jac_trivial]:
             with warnings.catch_warnings():
@@ -190,18 +348,13 @@ class BaseMixin(object):
 
     def test_full_result(self):
         res = least_squares(fun_trivial, 2.0, method=self.method)
-        # Use assert_almost_equal to check shapes of arrays too.
-        assert_almost_equal(res.x, np.array([0]), decimal=1)
-        assert_almost_equal(res.cost, 12.5)
-        assert_almost_equal(res.fun, np.array([5]))
-        assert_almost_equal(res.jac, np.array([[0.0]]), decimal=2)
-        # 'lm' works weird on this problem, thus only 3 digits.
-        assert_almost_equal(res.optimality, 0, decimal=3)
+        assert_equal(res.x, np.array([0]))
+        assert_equal(res.cost, 12.5)
+        assert_equal(res.fun, np.array([5]))
+        assert_equal(res.jac, np.array([[0]]))
+        assert_equal(res.optimality, 0)
         assert_equal(res.active_mask, np.array([0]))
-        if self.method == 'lm':
-            assert_(res.nfev < 25)
-        else:
-            assert_(res.nfev < 10)
+        assert_(res.nfev < 10)
         if self.method == 'lm':
             assert_(res.njev is None)
         else:
